🔒 chore(composer.json): update PHP version requirement to 8.1 The PHP version requirement has been updated to 8.1 to ensure compatibility with the latest version of PHP.
🚀 chore(composer.json): update Symfony dependencies to version 6.2.*
This commit updates the Symfony dependencies in the composer.json file to version 6.2.*. This ensures that the application is using the latest version of the Symfony framework, which includes new features, bug fixes, and security updates.

🚀 chore(composer.json): update dependencies and autoload
This commit updates the composer.json file to include the php-http/discovery plugin in the allow-plugins section. It also updates the symfony dependencies to version 6.2. Additionally, it adds an autoload-dev section to the composer.json file to include the App namespace in the src/ directory for development purposes.

🚀 feat(bundles.php): add WebProfilerBundle to all environments
🚀 feat(web_profiler.yaml): enable serializer data collection in the profiler
The WebProfilerBundle is now included in all environments, which allows for easier debugging and profiling of the application. Serializer data collection is now enabled in the profiler, which provides additional information about the serialized data.